Moved the application dir to the executable dir.
Having it in the temp folder caused issues where some files were sometimes removed if they weren't used since they were extracted, making parts of the app crash due to missing files on subsequent runs. This change avoids that, and having the folder next to the executable is nice for my uses anyway.